Initial commit of FPU division state machine.
Included are two different methods of testing the state machine.
The first is the file fp_div_tb.v which uses simulation to test
the division circuit. The second is an FPGA test harness for use
with the Digilent BASYS 3 board.

The comments in fpga_div.v document how to enter the rounding
attribute, dividend, and divisor into registers, and also how to
start execution of the state machine.